Changed request method from HEAD to GET
Was causing all server requests to fail on Gingerbread and above.
The code as it was SHOULD work, but this fixes the problem and doesn't
appear to increase response time. I don't believe it introduces any
performance penalty since the connection's request method is set
and then the response code is read, without content being transfered.